Deduplicate the local-day helper and rot-guard the inventory exclusions
lib/localDay.ts is now the single home of the local-midnight rule; the
task bucketing and conversation filtering modules import or re-export
from it instead of carrying byte-identical copies. Two older copies
(lib/voice/taskCounts.ts and a component) predate this PR and stay for a
follow-up rather than widening the diff.

The Windows inventory gains the macOS sibling's out-of-scope rot guard,
which immediately caught three prefixes matching no extracted route
(/v2/agent, /v1/config/api-keys, /v2/messages/); removed. The Flutter
inventory needs no exclusions at all: its streaming surfaces use other
bases, and the REST routes that do appear (including
/v2/messages/{id}/report and /v2/files) are modeled in the spec, so the
former blanket copy of the desktop list was excluding real in-scope
routes.

Verification: backend parity + inventories 19 passed; windows parity,
wire and conversations suites 110 passed; typecheck and eslint clean on
the touched files.
